api/requirements.txt
absl-py==0.15.0
asgiref==3.4.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
click==8.0.1
colorama==0.4.4
fastapi==0.68.1
flatbuffers==1.12
gast==0.4.0
google-api-core==2.1.1
google-auth==2.3.1
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.28.1
google-cloud-core==2.1.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.1
grpcio==1.41.1
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.8.1
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
nltk==3.6.5
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
proto-plus==1.19.6
protobuf==3.19.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.1
PySastrawi==1.2.0
python-dotenv==0.19.0
PyYAML==5.4.1
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.25
starlette==0.14.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
absl-py==0.15.0
asgiref==3.4.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
click==8.0.1
colorama==0.4.4
fastapi==0.68.1
flatbuffers==1.12
gast==0.4.0
google-api-core==2.1.1
google-auth==2.3.1
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.28.1
google-cloud-core==2.1.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.1
grpcio==1.41.1
gunicorn==20.1.0
h11==0.12.0
h5py==3.1.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.8.1
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
nltk==3.6.5
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
proto-plus==1.19.6
protobuf==3.19.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.1
PySastrawi==1.2.0
python-dotenv==0.19.0
PyYAML==5.4.1
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.25
starlette==0.14.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
